Move `ContractAbi` package
