+++ src/common/blockchain/index.js
@@ -1,6 +1,7 @@
 import * as Ethereum from './ethereum'
 import * as Plasma from './plasma'
 import * as Contract from './contract'
+import * as ContractABI from './contractAbi'
 import * as GasEstimator from './gasEstimator'
 import * as TxOptions from './txOptions'
 import * as Token from './token'
@@ -11,6 +12,7 @@ export {
   Ethereum,
   Plasma,
   Contract,
+  ContractABI,
   GasEstimator,
   TxOptions,
   BlockchainDataFormatter,
