+++ src/common/services/ethereumService.js
@@ -1,6 +1,6 @@
-import { Ethereum, Token } from 'common/blockchain'
+import { Ethereum, Token, ContractABI } from 'common/blockchain'
 import { ContractAddress } from 'common/constants'
-import { ContractABI, Datetime, Parser } from 'common/utils'
+import { Datetime, Parser } from 'common/utils'
 import { providerService } from 'common/services'
 
 export const fetchAssets = async (provider, address, lastBlockNumber) => {
